backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
django-cors-headers>=4.3
python-decouple>=3.8
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
# Dev / Quality
black
flake8
isort
pytest
pytest-django
